--- EmailConstants.java
@@ -1,3 +1,0 @@
-public interface EmailConstants {
-    public static final int PORT = 56789;
-}
