+++ src/EmailConstants.java
@@ -1,0 +1,3 @@
+public interface EmailConstants {
+    public static final int PORT = 56789;
+}
