+++ src/SMTP.java
@@ -1,0 +1,21 @@
+
+/**
+ * NO LONGER USED
+public class SMTP {
+
+
+    public SMTP(){
+
+    }
+
+    public void servStartCon(){ // change void to String
+
+        //return "220 " + client.getInetAddress().getHostAddress()
+    }
+
+    public void clientStartCon(){ // change void to String
+
+        // return "HELO " + 
+    }
+}
+*/
